Attributes are removed from group classes.
Attributes read in the input mesh are converted to group names so as
to be renamable by the user and listed in the meshe's groups, but
additional mesh family items may be built when group names are convertible
to integers so that user subroutines looping on families and family
properties still work as before.

Also, selectors still contain internal attributes so as to ensure a
range[] of integers works as expected, but these are now built directly
from group names which are convertible to integers.